gpu: single CPU/GPU Docker image, merged server script, 64-bit MC paths
Three related changes that make the mcl_gpu engine usable in production:

- mcl: store the Monte-Carlo path count as a 64-bit `long` end to end
  (config field, new YamlConfig::GetLong getter, object_registry read,
  cluster path-split, CPU sweep loop counters). Previously `paths` was an
  `int` and GetInteger truncated at parse time, so any value above 2^31
  (e.g. 10e9) wrapped negative and was rejected with "paths must be > 0".
  Verified on the GPU: 10e9 paths price the ATM call at 15.71132 (analytic
  15.71131), std error scaling textbook 1/sqrt(N) across 1M..10e9.

- docker: merge Dockerfile + Dockerfile.gpu into one parameterised
  Dockerfile (BUILD_BASE / RUNTIME_BASE / ENABLE_CUDA / CUDA_ARCH build
  args); the GPU variant just swaps in the nvidia/cuda base images. Fixes
  the stale CUDA tag (12.4.1 has no ubuntu24.04 image -> 12.6.3, still
  sm_89, same libcudart.so.12). build-gpu/ added to .gitignore.

- docker: fold run_docker_server_gpu.sh into run_docker_server.sh behind a
  `--gpu` flag (with --arch); the shared thoth_build_image helper now builds
  either variant. CPU run path is unchanged.

- devcontainer: install CUDA (nvcc + cudart-dev + curand-dev) via apt in the
  Dockerfile instead of the nvidia-cuda feature, which pulls libtinfo5 and
  breaks on Ubuntu 24.04.

README, docs/gpu_mcl.md, CMakeLists comment and samples/gpu_call.yaml synced.

+++ src/core/object_registry.cpp
@@ -433,7 +433,7 @@ map<string, ObjectManager::Factory> MakeRegistry()
         MclConfiguration* M = m.collector().Add( std::make_unique<MclConfiguration>( n ) );
         M->_max_time_step = m.cfg().GetInteger( n + ".max_time_step" );
         M->_min_time_step = m.cfg().GetInteger( n + ".min_time_step" );
-        M->_paths = m.cfg().GetInteger( n + ".paths" );
+        M->_paths = m.cfg().GetLong( n + ".paths" );
         //! year-fraction sub-step, so it is a double (0.01, not 0)
         M->_vol_time_step = m.cfg().GetDouble( n + ".vol_time_step" );
         M->_node_file = m.cfg().GetString( n + ".node_file", MCL_NODE_PATH );
